requirements.txt
Django>=4.0,<5.0
mysqlclient==2.1.1
djangorestframework==3.14.0
django-cors-headers==3.13.0
python-decouple==3.6
